+++ packages/docs/src/vite-plugin-package-versions/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-param-reassign */
 import { readFileSync } from 'fs';
 import { join } from 'path';
 import { Plugin } from 'vite';
@@ -8,15 +9,13 @@ interface PackageVersionsPluginOptions {
 
 function VitePluginPackageVersions(options: PackageVersionsPluginOptions): Plugin {
   return {
-    name: 'vite-plugin-package-versions',
     configResolved(config) {
       const versions = options.packagePaths.reduce((acc, packagePath) => {
         try {
           const packageJson = JSON.parse(
             readFileSync(join(config.root, packagePath, 'package.json'), 'utf-8'),
-          );
+          ) as { name: string, version: string };
           acc[packageJson.name] = packageJson.version;
-          console.log(acc);
         } catch (error) {
           console.error(`Failed to read package.json from ${packagePath}:`, error);
         }
@@ -29,6 +28,7 @@ function VitePluginPackageVersions(options: PackageVersionsPluginOptions): Plugi
         __VITE_PACKAGE_VERSIONS__: JSON.stringify(versions),
       };
     },
+    name: 'vite-plugin-package-versions',
   };
 }
